Complex API (#302)
Allow the user to discretize complex `SesquilinearForm`, `LinearForm`
and `Functional` and to assemble them. For this, the user needs to add
to the `SymbolicSpace` of SymPDE an attribute named
`codomain_type='complex'`.

Also, add as a comment the case where the user gives a
`SesquilinearForm` (not yet implemented in SymPDE), which will be
discretized into an object of type `DiscreteSesquilinearForm`.

The different projectors can now support complex values.

Adds some tests for the complex assembly in `test_assembly.py`, some
tests to check the precision of the solution in 'test_complex.py' for a
domain with and without mapping, both single- and multi-patch.

Add `domain_dtype` property to `DefNode`, to build the correct decorator
for constant in Pyccel. Also, this parameter corrects the importation of
mathematical function in the complex case from `math` to `numpy`, thus
the function can handle the complex values.

NOTE: from now on when declaring a SymPDE `Constant` for Psydac, the
type of the constant must be specified with one of the boolean flags
`complex`, `float`, or `integer`.

Add `dtype`property to :
- MatrixGlobalBasis
- MatrixLocalBasis
- MatrixQuadrature
- BlockStencilMatrixLocalBasis
- BlockStencilMatrixGlobalBasis
- BlockStencilVectorLocalBasis
- BlockScalarLocalBasis
- StencilMatrixLocalBasis
- StencilVectorLocalBasis
- StencilVectorGlobalBasis
- EvalField
- LocalElementBasis
- GlobalElementBasis
- TensorIterator
- MatrixRankFromCoords
- LocalElementBasis
- ScalarLocalBasis
- ComputeLogicalBasis
- GeometryExpr
- GeometryExpressions
- Zeros

Add the `prefix` property to : 
- BasisAtom
- IteratorBase
- LogicalBasisValue
- Loop

Add `change_dtype` function, which copies the `vector_space`of a
`FemSpace` and changes its data type.

---------

+++ psydac/linalg/stencil.py
@@ -123,6 +123,7 @@ class StencilVectorSpace( VectorSpace ):
         if cart.is_parallel and not cart.is_comm_null:
             self._mpi_type      = find_mpi_type(dtype)
             if isinstance(cart, InterfaceCartDecomposition):
+                # TODO : Check if this line really change the ._shape
                 self._shape = cart.get_interface_communication_infos(cart.axis)['gbuf_recv_shape'][0]
             else:
                 self._synchronizer = get_data_exchanger( cart, dtype , assembly=True, blocking=False)
@@ -248,7 +249,7 @@ class StencilVectorSpace( VectorSpace ):
         if cart.is_comm_null: return
         if isinstance(cart, InterfaceCartDecomposition):
             # Case where the patches that share the interface are owned by different intra-communicators
-            space = StencilVectorSpace(cart)
+            space = StencilVectorSpace(cart, dtype=self.dtype)
             self._interfaces[axis, ext] = space
         else:
             # Case where the patches that share the interface are owned by the same intra-communicator
@@ -271,6 +272,8 @@ class StencilVectorSpace( VectorSpace ):
                     parent_ends[axis] = self.pads[axis]
 
             cart = cart.change_starts_ends(tuple(starts), tuple(ends), tuple(parent_starts), tuple(parent_ends))
+
+            #TODO Check if we create object from it, otherwise its only purpose is to store some parameters which is innefficient
             space = StencilVectorSpace(cart, self.dtype)
 
             self._interfaces[axis, ext] = space
@@ -2504,7 +2507,8 @@ class StencilInterfaceMatrix(LinearOperator):
                                     flip_axis=self._flip,
                                     interface_axis=self._codomain_axis,
                                     d_start=(self._domain_start,),
-                                    c_start=(self._codomain_start,))
+                                    c_start=(self._codomain_start,),
+                                    dtype= self.dtype)
 
                     starts = self._args.pop('starts')
                     nrows  = self._args.pop('nrows')
@@ -2530,7 +2534,8 @@ class StencilInterfaceMatrix(LinearOperator):
                                             flip_axis=self._flip,
                                             interface_axis=self._codomain_axis,
                                             d_start=(self._domain_start,),
-                                            c_start=(self._codomain_start,))
+                                            c_start=(self._codomain_start,),
+                                            dtype= self.dtype)
 
                     starts      = self._args.pop('starts')
                     nrows       = self._args.pop('nrows')
@@ -2564,7 +2569,8 @@ class StencilInterfaceMatrix(LinearOperator):
                                         flip_axis=self._flip,
                                         interface_axis=self._codomain_axis,
                                         d_start=(self._domain_start,),
-                                        c_start=(self._codomain_start,))
+                                        c_start=(self._codomain_start,),
+                                        dtype= self.dtype)
 
                 self._args = {}
